requirements.txt
chromadb>=0.5.0
langgraph>=0.2.0
pypdf>=4.2.0
typing_extensions>=4.10.0
openai>=1.40.0
python-dotenv>=1.0.1
fastapi>=0.110.0
jinja2>=3.1.0
